rename loadPartyDetailsCurrent.js to loadPartyDetails.js to be consistent with function name; move to src directory
